moved presets out of oi, created preset class, made respective changes
