+++ src/main/java/frc/robot/OI.java
@@ -17,7 +17,6 @@ import frc.robot.lib.DPADButton;
 import frc.robot.lib.Gamepad;
 import frc.robot.lib.LogitechController;
 import frc.robot.lib.Xbox;
-import frc.robot.subsystems.Wrist.Preset;
 import frc.robot.commands.*;
 
 /**
@@ -58,10 +57,6 @@ public class OI {
     XBOX,         // 0
     DUALSHOCK     // 1
   }
-
-  /* TODO: add more presets as we need them */
-  public static Preset[] wristPresets = new Preset[]{Preset.BOTTOM, Preset.TOP};
-  public int currentPreset = 1;
 
   /*
    * OI-related variable naming conventions:
@@ -127,8 +122,8 @@ public class OI {
     //d_btn_LT.whileHeld(new IntakeCargo());
       
     /* TODO: Add o_operaterOI button commands */
-    this.o_btn_RB.whenPressed(new PresetWristControl(nextPreset()));
-    this.o_btn_LB.whenPressed(new PresetWristControl(previousPreset()));
+    this.o_btn_RB.whenPressed(new PresetWristControl(Robot.m_wrist.nextPreset()));
+    this.o_btn_LB.whenPressed(new PresetWristControl(Robot.m_wrist.previousPreset()));
   }
 
   /* Used by the DriveTrain subsystem for default command */
@@ -138,23 +133,5 @@ public class OI {
 
   public LogitechController getOperatorOI() {
     return this.o_operatorOI;
-  }
-
-  public Preset nextPreset() {
-    if(currentPreset == wristPresets.length-1) {
-      return wristPresets[currentPreset];
-    } else {
-      currentPreset += 1;
-      return wristPresets[currentPreset];
-    }
-  }
-
-  public Preset previousPreset() {
-    if(currentPreset == 0) {
-      return wristPresets[currentPreset];
-    } else {
-      currentPreset -= 1;
-      return wristPresets[currentPreset];
-    }
   }
 }
