+++ src/main/java/frc/robot/commands/PresetWristControl.java
@@ -9,7 +9,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
-import frc.robot.subsystems.Wrist.Preset;
+import frc.robot.util.Preset;
 
 public class PresetWristControl extends Command {
   private Preset preset;
