+++ src/main/java/frc/robot/util/Preset.java
@@ -1,0 +1,23 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
+
+package frc.robot.util;
+
+/**
+ * Add your docs here.
+ */
+public class Preset {
+    private final double pos;
+
+    public Preset(double pos) {
+        this.pos = pos;
+    }
+
+    public double getPos() {
+        return pos;
+    }
+}
